add do_pred flag to allow for eval-only ow

+++ control_simp/models/eval.py
@@ -89,7 +89,7 @@ def run_evaluation(df, x_col="complex", y_col="simple", pred_col="pred", metrics
 
 class Launcher(object):
 
-    def bart(self, model_loc, test_file, out_dir, name, ctrl_toks=None, max_samples=None, samsa=True, device="cuda", ow=False, num_workers=8, mtl=False, beams=10):
+    def bart(self, model_loc, test_file, out_dir, name, ctrl_toks=None, max_samples=None, samsa=True, do_pred=True, device="cuda", ow=False, num_workers=8, mtl=False, beams=10):
         start = time.time()
 
         pred_file = f"{out_dir}/{name}_preds.csv"
@@ -104,7 +104,7 @@ class Launcher(object):
         model = BartFinetuner.load_from_checkpoint(model_loc, mtl=mtl, strict=False).to(device).eval()
 
         # run generation on test data
-        if ow or not os.path.isfile(pred_file):
+        if do_pred and (ow or not os.path.isfile(pred_file)):
             print("Generating predictions...")
             test_set["pred"] = run_generator(model, test_set, ctrl_toks=ctrl_toks, max_samples=max_samples, num_workers=num_workers, beams=beams)
             test_set.to_csv(pred_file, index=False)
